requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
rich==13.7.0
google-generativeai==0.3.2 